Fix: let fetch errors bubble up
Fixes #3

Rather than letting the errors from a call to the fetch method get
swallowed, we let them throw and let the user decide how to handle them.
The code that was there previously seems handy for development, but is
atypical for production.

This also leads to more reliable typing by avoiding returning an empty
object in the error case.

I don't love copying types out of the `@types/tiny-json-http` module,
but it doesn't expose the types we need and due to type overloading, I
wasn't able to extract them with tricks like `Parameters` or
`ReturnType`.

A cleaner fix would be to modify that module to export what we need, but
I've only got so many hours in the day ;)